requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
joblib==1.3.2
numpy==1.26.4
pandas==2.2.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.3.0
typing_extensions==4.9.0
tzdata==2024.1
gunicorn==21.2.0